Revert "Added a 'to_world' transform to all shape classes"
This reverts commit 73abb5436db0e0dfafc73ee412cc37372183d0b8.
