+++ docs/src/release_notes/v0.25.x.md
@@ -1,28 +1,4 @@
-# v0.25.x series (current stable)
-
-## v0.25.1 (upcoming release)
-
-% ### Deprecated
-%
-% ### Removed
-%
-
-### Added
-
-* Split bitmap-to-xarray converter between data array and dataset steps
-  ({ghpr}`383`).
-* The {class}`.Shape` now supports a `to_world` transform, making it a available
-  for all child classes. Since the {class}`.Buffermesh` does not support this
-  kind of transform, a helper function is added to the `warp.py` module and
-  a tutorial to explain its use is available.
-* Added an Ubuntu-Based Dev Container for the Eradiate development environment
-  ({ghpr}`384`).
-
-% ### Changed
-%
-% ### Fixed
-%
-% ### Internal changes
+# v0.25.x series
 
 ## v0.25.0 (20th December 2023)
