+++ docs/src/release_notes/v0.27.x.md
@@ -22,8 +22,6 @@
   facilitate absorption database downloads ({ghpr}`397`).
 * Gaussian SRF datasets can now be dynamically generated using the
   :func:`.srf_tools.make_gaussian` function ({ghpr}`401`).
-* All `.Shape` classes now support a `to_world` member, which defines an arbitrary
-  transformation of the object.
 
 ### Changed
