+++ src/eradiate/scenes/shapes/_filemesh.py
@@ -54,12 +54,9 @@ class FileMeshShape(ShapeNode):
             raise ValueError(
                 f"unsupported mesh file extension '{self.filename.suffix}'"
             )
-        result = {
+
+        return {
             "type": mi_plugin,
             "filename": str(self.filename),
             "face_normals": True,
         }
-        if self.to_world is not None:
-            result["to_world"] = self.to_world
-
-        return result
